Refactor error handling; fix cursor pagination; update button submit behavior

+++ client/src/components/nav/Invite.jsx
@@ -2,12 +2,15 @@ import { useState } from "react";
 import apiFetch from "../../api/client";
 import { useChats } from "../../context/ChatProvider";
 
-export default function Invite({ i }) {
+export default function Invite({ i, setError }) {
   const { setRefreshTrigger } = useChats();
-  const [error, setError] = useState();
+  const [isSubmit, setIsSubmit] = useState(false)
   const [isActive, setIsActive] = useState(true);
 
   const handleAccept = async () => {
+    if(isSubmit) return;
+    setError()
+    setIsSubmit(true);
     try {
       const options = {
         method: 'POST',
@@ -18,10 +21,15 @@ export default function Invite({ i }) {
       setRefreshTrigger(prev => prev + 1)
     } catch (error) {
       setError(error)
+    } finally {
+      setIsSubmit(false)
     }
   }
 
   const handleDecline = async () => {
+    if(isSubmit) return;
+    setError()
+    setIsSubmit(true);
     try {
       const options = {
         method: 'DELETE',
@@ -31,6 +39,8 @@ export default function Invite({ i }) {
       setIsActive(false);
     } catch (error) {
       setError(error)
+    } finally {
+      setIsSubmit(false)
     }
   }
   
@@ -41,7 +51,6 @@ export default function Invite({ i }) {
           {i.avatarUrl ? <img src={i.avatarUrl}></img> : <div>{i.name[0]}</div>}
           <div>
             <p>{i.name}</p>
-            <p>{error}</p>
           </div>
           <button onClick={handleAccept}>Accept</button>
           <button onClick={handleDecline}>X</button>
